rpc: factor out logic for getting NodeInfo
