+++ crates/sui-rpc-api/src/client/sdk.rs
@@ -27,7 +27,6 @@ use crate::rest::checkpoints::CheckpointResponse;
 use crate::rest::checkpoints::ListCheckpointsQueryParameters;
 use crate::rest::coins::CoinInfo;
 use crate::rest::health::Threshold;
-use crate::rest::info::NodeInfo;
 use crate::rest::objects::DynamicFieldInfo;
 use crate::rest::objects::ListDynamicFieldsQueryParameters;
 use crate::rest::system::GasInfo;
@@ -41,6 +40,7 @@ use crate::rest::transactions::ResolveTransactionResponse;
 use crate::rest::transactions::TransactionExecutionResponse;
 use crate::rest::transactions::TransactionResponse;
 use crate::rest::transactions::TransactionSimulationResponse;
+use crate::types::NodeInfo;
 use crate::types::X_SUI_CHAIN;
 use crate::types::X_SUI_CHAIN_ID;
 use crate::types::X_SUI_CHECKPOINT_HEIGHT;
